[omnibox] Moves ranking of local history zero-prefix suggestions to C++
In order to support different ranking approaches for the local history
zero-prefix suggestions, this CL modifies the current implementation to
allow ranking the normalized search terms in C++.

To accomplish this, LocalHistoryZeroSuggestProvider now requests an
unordered list of rows grouped by |normalized_term| from the URL
database in which |visit_count| is the sum of all visit counts and
|last_visit_time| is the latest visit time for that normalized term.

It also simplifies the LocalHistoryZeroSuggestProvider implementation by
moving the freshness and validity checks for the normalized term into
the SQL query.

This change is not expected to alter the existing behavior as confirmed
by the unit tests.

Bug: 1096615
Change-Id: I01c29afff51f348e5c32766e4e53b3ff7687a3b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278625
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786019}