arc: Add native_bridge_64bit_support_experiment
Add an experiment flag to enable 64-bit native bridge for ARC on
systems that have 64-bit native bridge support available but not
enabled.

Note that due to the nature of the timing of starting ARC mini-
container - which can be affected by the properties controlled by
this experiment - and user login, the experiment is treated as a one
way off->on transition for all users on a given device.

The experiment flag will have no effect on systems that do not have
64-bit native support available, nor will it affect systems that
have 64-bit native support enabled generally.

Bug: b:62095998
Change-Id: If2a5dc96ba36f8aba31405122a6de20764af30b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2252956
Commit-Queue: Josh Horwich <jhorwich@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786022}