DevTools: protocol and backend support for grid areas
This introduces a new option (showAreaNames) to the
gridHighlightConfig option, passed to overlays.
The option defaults to false. When set to true, it tells the
backend implementation of the overlay domain to
also generate path information for highlighting grid
areas in the page.

Bug: 1100384
Change-Id: I75380034d2363f06f1987c6fb5a4cf17c38eb68f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275484
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#786034}