Fix a crash when tablet mode ends before dropping a window
When a window is dragged to overview by gesture dragging from
shelf, tablet mode can end before the drag ends, at which point
The homescreen window is closed, and we can't use it to get the
SplitViewController.

This CL reworks ShouldShowHomeScreen() to fix the crash, and do only
the next checks if the previous ones failed.

BUG=1101143
TEST=Manually by exiting tablet mode while dragging a window from shelf.

Change-Id: I6a58e1c7caa6ebccc45d47cbfa27136891cfee9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2285609
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786082}