Remove vr_testapp from bots
The tool currently fails to compile, and is breaking the trybots.
Normally, we'd revert the commit that caused this to fail, but this tool
is scheduled for deletion in the very near future and currently unused;
so fix the bots by stopping building it on them.

Change-Id: Ic4e8639c3d2f3f083072dd21ffa8b3d5c7df70ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2285225
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786113}