Migrate timers for off hours and session length to a wall clock
Policy implementations for OffHours and SessionLengthLimit use
OneShotTimer for measuring time. This timer uses clock like device's
uptime and therefore is paused when device is suspended, which leads to
the problem: if device was suspended during the session, session length
limit is not applied correctly, the same may happen with off hours.

This commit changes timers used for these policies to a WallClockTimer,
which takes possibility of suspend/resume into account and binds to the
"real" time. As part of it, this commit changes TimeTicks to Time in
numerous of places, notable one is the session.start_time pref, which
now will store internal representation of Time, not TimeTicks.

Bug: 1095645
Change-Id: Ia4950561f695053091b36f1ed481807c6d0ed6a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274395
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Commit-Queue: Oleg Davydov <burunduk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786248}