[b4p/sign-out] Reduce expectations of sign-out bubble integration test
This is a preparation to allow the user to choose which passwords to
save in the 'save unsynced credentials' bubble. To do so, we will
modify ManagePasswordsUIController::SaveUnsyncedCredentials() to take
the vector of credentials as a parameter.

BubbleObserver::AcceptSaveUnsyncedCredentialsPrompt() fakes the user
accepting the bubble in this integration test. If it is kept in the new
design, it would need to take the credentials as a parameter as well,
which renders some expectations in this test completely artificial.

So, after this CL, the test no longer verifies that accepting the bubble
stores the credentials in the profile store. Instead it only verifies
that the bubble is opened. The saving logic is still covered by unit
tests in ManagePasswordsUIController.

Bug: 1098672
Change-Id: I7a1d752b67934e31d968ac23475026fc0cdc4af8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2287315
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786276}