[ozone] Refactored drag and drop
Running the drag loop in DesktopDragDropClientOzone::StartDragDrop
caused issues in the X11 platform implementation. The platform needs to
capture and dispatch events properly, which is typically handled by
|X11WholeScreenMoveLoop| that runs its own run loop.

This CL moves the drag loop to the platform level.  Two existing Ozone
platforms are affected: changes to Wayland are trivial, and X11 gets
some code that duplicates a few bits of |DesktopDragDropClientAuraX11|.
This is temporary and will be fixed soon when we finally obsolete the
non-Ozone path and get rid of |DesktopDragDropClientAuraX11|.

Also, a few minor things are fixed that had been proposed earier in CLs
based on the previous approach:
1. The drag image is now created only on platforms that need it [1]
2. The ui::XDragDropClient is documented [2]

[1] https://crrev.com/c/2264553
[2] https://crrev.com/c/2262612

Bug: 1014860
Change-Id: Iec8aa145a23e1002df06a788bbc8ecb73d04ec63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2267378
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Nick Yamane <nickdiego@igalia.com>
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Cr-Commit-Position: refs/heads/master@{#786324}