Spare unnecessary copies in NotifyUnsyncedCredentialsWillBeDeleted()
The sole purpose of these APIs is to propagate the unsynced credentials
computed in PasswordSyncBridge up to ManagePasswordsState, which will
then own them until the user decides to save or discard. So this CL
modifies the APIs to a pass-by-value/move scheme. This spares 2 copies:
when posting the task to UnsyncedCredentialsDeletionNotifier; and when
storing the credentials in ManagePasswordsState.

Bug: None
Change-Id: Ic2f0dc191700b55d74b199b130a55d71f61150a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283751
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#786326}