[CBCM] Create factory for report generation delegates
Rather than having parent objects pass concrete delegates to their
members, the idea of this change is to have one factory object that will
create platform-specific delegates for all the enterprise reporting
classes. That way, only the factory needs to be passed around, and each
reporting class can use it to get the delegate they need. This helps
encapsulation because delegates of parent objects don't need to know
about their children and their delegates.

For now, the concrete factory's life cycle is owned by ReportGenerator
because BrowserReportGenerator is the only one using it so far. But
over the course of the next few CLs, the factory will be owned by
objects higher and higher in the ownership hierarchy, until it's owned
by CBCMController.

Bug: 1091916
Change-Id: I7eec88ab1b1fb1bbc4c24ab198cc91637b836093
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2281031
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786329}