Revert "device/fido: only send the single matching credential ID."
This reverts commit c6728c245d96202d24caa5531ee2f5ec7d8c275a.

The credential is optional in the response when there's a single entry in the allowlist. Normally we wouldn't be sending silent requests in that case, but it can happen if the allow list had more entries before filtering or if there's an appid extension.

This is a clean revert for merging purposes. Test changes and a fix will come later.

BUG=1102777

Original change's description:
> device/fido: only send the single matching credential ID.
> 
> When prelighting credential IDs for getting an assertion, only send the
> matching credential ID on the following up=1 request, rather than the
> whole matching batch.
> 
> Change-Id: I41534e82306b82c2332b44deb6466a9ee46c2905
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250893
> Commit-Queue: Adam Langley <agl@chromium.org>
> Reviewed-by: Martin Kreichgauer <martinkr@google.com>
> Cr-Commit-Position: refs/heads/master@{#780525}

TBR=agl@chromium.org,martinkr@google.com

# Not skipping CQ checks because original CL landed > 1 day ago.

Change-Id: I3779623933af956f2cd8ade9200d7e7165cc55fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2287384
Reviewed-by: Adam Langley <agl@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Adam Langley <agl@chromium.org>
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Auto-Submit: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786332}