Determine the context of subframe-initiated subresource requests.
We wish to know whether a request to which variations headers will be
appended is from a frame whose top-level domain is a Google-owned web
property, e.g. YouTube, DoubleClick, etc.

In order to determine this context for a subframe-initiated subresource
request on the renderer side (e.g. a request to DoubleClick from an ad
on foo.com), we cannot use ResourceRequest::TrustedParams because they
are not populated. It is, however, possible to determine the context in
the renderer and to pass it along to the code that appends headers.

Bug: 1094303
Change-Id: I6d62b1ba52425a9b0d80d56b20326e23b6949263
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276231
Commit-Queue: Caitlin Fischer <caitlinfischer@google.com>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786349}