Create child frame views with copy of parent's ScreenInfo.
When a RenderWidgetHostViewChildFrame is created, any calls to
GetScreenInfo() will default to getting the default ScreenInfo if the
view's FrameDelegate hasn't been attached yet. This can lead to
incorrect sizing if the view is on a secondary monitor with a different
device scale factor.

But at time of the view's creation we do know the parent view, so we
can give the child view a copy of it on creation. This CL allows an
initial ScreenInfo to be set on the RenderWidgetHostViewChildFrame at
creation time.

Bug: 966852
Change-Id: Ie80c617d8c6c741c33cfbddfdd63282ff28b1e9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283404
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786359}