Enable SurfaceLayer video with VizForWebView
With VizForWebView enabled webview now supports embedding of the
surfaces. Unfortunately because of the order how we apply features
it's not possible to override DisableSurfaceLayerForVideo from the finch
config.

This CL modifies the check for disabling surface layer to take into
account VizForWebView experiment.

Bug: 1103197
Change-Id: Ied5f9e85680e0fa151c79020a5b8de7ed92f2812
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2287809
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786436}