Prevent HungRendererDialogView from appearing during deep scans
This dialog can appear during a long scan, or if the user waits for too
long to close the scan dialog indicating the scanning verdict. It is
misleading as the page seems unresponsive because of the scanning, so
this CL prevents that dialog from appearing at the same time as the
scanning dialog.

The dialog is prevented from being shown by having
ShouldIgnoreUnresponsiveRenderer return true until the paste callback
has been ran.

Bug: 1028750
Change-Id: If34f8184f254074c9ce30b49337a14531e24a8c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283397
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Roger Tawa <rogerta@chromium.org>
Commit-Queue: Dominique Fauteux-Chapleau <domfc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786451}