[Tab Groups Collapse] Fix drag for collapsed groups.
The bug that makes it not possible to drag a tab to the right of a
collapsed group header is due to TabStrip::GetLastVisibleTab() clamping
the insertion index.

TabStrip::ShouldTabBeVisible is used for determining if a tab is visible
in the tabstrip due to space constraints. Moving the logic for collapsed
tabs into TabStrip::SetTabVisibility().

This should also make is slightly nicer and show less headers in the case
where the tabstrip is full of group headers and not visible tabs.

Bug: 1099367
Change-Id: If31bf23b9d2fc069481f092e38310635b89d3165
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2277381
Commit-Queue: Charlene Yan <cyan@chromium.org>
Reviewed-by: Connie Wan <connily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786498}