Link Doctor Removal Part 2: Tracking Requests.
This CL removes the Link Doctor click tracking network requests and UMA
code. It also removes the code promote the Google cache suggestion and
record when the cache suggestions are interacted with.

This CL also renames the variable "src" to "source" in the Javascript
files, as otherwise, removing the trackClick method in the JS file
results in funky build failures on Android, likely due to a buggy regexp
when trying to inline files.

Bug: 1030281
Change-Id: Ia2a9f56397f1014baf65ac0374c446a062b2007d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2284973
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Edward Jung (EMEA) <edwardjung@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786511}