Batch ContextualSearchManagerTests
This change introduces test batching to ContextualSearchManagerTest.
This means that the browser process (and Instrumentation instance) is
not restarted between each of the tests in the suite. However, it is
restarted before and after this suite, as the batching type is
PER_CLASS.

For ContextualSearchManagerTest, this means the tests run >5x faster.

I had to modify a few tests to clean up persistent state changes, and I
got rid of the 1 second delay added to each test, replacing it with
retries when panel interactions fail.

I also had to make the EmbeddedTestServerRule reusable so that the
server can be restarted for each test. In the future this should be
further improved to allow the test server to be reused across tests.

Finally, I had to remove a maximizePanelThenPromoteToTab animation from
ContextualSearchManager that wasn't visible anyways (as it is used
when launching an external app), as when run in the test it wasn't being
set up correctly and ended up being too invasive to tear down.

Bug: 989569
Change-Id: I5aff1667c01ac3dceb33704c6853299cab1cd541
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270713
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Donn Denman <donnd@chromium.org>
Reviewed-by: Matt Mueller <mattm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786516}