Make simplified domain animation smoother by fading
This CL adds in fading to/from transparent alongside the existing
elision animation to/from simplified domain. It's similar to the
previous PathFadeAnimation that faded out the path except that it (a)
happens concurrently with changing the URL's display bounds, and (b)
it can apply to subdomains as well as the path.

Bug: 1090393
Change-Id: I8f9348dc47dce3260969577feb5517b14483d558
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2285074
Commit-Queue: Emily Stark <estark@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786529}