FormControlsRefresh is now the default - move tests out of virtual [4]
This is one CL of many, moving tests out of the virtual/controls-refresh
directory, into fast/forms. This CL moves:

 virtual/controls-refresh/focus-rect -> fast/forms/focus-rect

This is a pure move of the above folder(s). A few tests share names
in the old/new location, but in all cases, the version being
moved is updated/new and should overwrite the old version.

See also crbug.com/1102647 - there is additional cleanup that will
be needed after this move is complete. But to keep the review process
simple, this is just a pure move.

Bug: 1034611
Change-Id: I99e246142240ba872d8b3a0bec0815769634aaf6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2284020
Auto-Submit: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786547}