Android: CompositorViewHolder manages viewport sizing
This CL moves the logic in ChromeFullscreenManager that controls
viewport to CompositorViewHolder so that it will be the source of
viewport information.

ContentView sizing/offset update is still initiated by
ChromeFullscreenManager but they are decoupled(i.e. no more
ViewSizeDelegate), and now the flow is handled through CVH
listening to BrowserControlsStateProvider.onControlOffsetChanged.

Bug: 966272
Change-Id: I111b47d4fd93f4d1866dc0cab7012c9339278788
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278742
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786580}