device/fido: fork requests early in MakeCredentialRequestHandler.
The |CtapMakeCredentialRequest| objects are already tweaked for each
authenticator, but it was done just before dispatch. Thus lots of logic
depended on the authenticator-independent |request_|, which wasn't what
was actually sent. Since the tweaks are currently small, that didn't
matter, but hmac_secret will make that more complex.

Instead, perform the tweaks at the beginning of processing and carry
that copy of the request through the functions. The logic becomes
authenticator-specific after a while, so it would be possible to use
|request_| for this. But, in order to avoid having to remember which
functions are request specific and which can use |request_|, just thread
the object everywhere.

Since the |CtapMakeCredentialRequest| is quite a complex object, with a
lot of members to move, the per-authenticator request is held in a
|unique_ptr| so that the moving is trivial.

Change-Id: I807aa44df237440c209d3062621b66e953515854
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2258855
Commit-Queue: Adam Langley <agl@chromium.org>
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#786593}