Per-desk Shelf application menus
When there are multiple running instances of the same
app, left clicking its icon on the shelf will display
an application menu showing an item for each instance
from which users can select which instance to activate.

When the kPerDeskShelf feature in enabled, this list
should be filtered to show only running instances in
the current active desk. This is done via giving the
ShelfItemDelegate a filter_predicate to exclude the
apps belonging to other desks.

BUG=997102
TEST=Manually, browser_tests --gtest_filter=*PerDeskShelfAppBrowserTest.AppMenus*

Change-Id: I8d70586a1ad638a42e6aa71378a55007c6b8dabb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2258902
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786594}