Remove cr-input from CrOS destination dropdown
cr-input is mainly used for receiving user keyboard input and comes
with a lot of built-in features to support that. Since the print
preview dropdown does not directly receive keyboard input it makes
sense to replace cr-input with a standard div. This makes customization
of the dropdown cleaner as well.

"input" was removed from all of the element id's to reflect the removal
of cr-input.

Bug: 1059607
Change-Id: Icdbde1ac3e089fc76b35eef2ba67f76a4f649fe3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2284866
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Bailey Berro <baileyberro@chromium.org>
Commit-Queue: Gavin Williams <gavinwill@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786604}