Avoid address-space exhaustion on 32-bit test processes
When running base_perftests for x86 an out-of-memory failure is hit on
MemoryAllocationPerfTest.SingleBucket/3. This test was already disabled
on Android due to its excessive use of memory and this change disables
the test for all 32-bit processes.

With this change it is now possible to run base_perftests for x86.

Change-Id: If3cfbd8b754ae6f09f70e989ac1cab94e2a188be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2288829
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786633}