[Translate] Componentize translate_utils.*
For WebLayer's initial translate infobar implementation in 85 we forked
Chrome for Android's translate infobar impl as we didn't know what
changes we would need to make and how much of it would be relevant. That
initial implementation has now landed and we have a much clearer
picture, so it makes sense to start sharing the common infrastructure.
This CL starts that process by componentizing //chrome's
translate_utils.*. We adjust namespaces accordingly (including of
generated Java enums) and remove an unused enum in the process.

Bug: 1101240
Change-Id: If094979c94e65fdcc3d311cb7c39a03d76ff71f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276197
Reviewed-by: Megan Jablonski <megjablon@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786671}