[b4p/sign-out] Pass selected credentials to save method
No behavior is changed. This CL starts storing the unsynced credentials
in the bubble controller and modifies the signature of
ManagePasswordsUIController::SaveUnsyncedCredentials() to receive a
vector of selected_credentials as a parameter. For the moment, all the
credentials are passed when the user clicks Save. Future CLs will allow
selecting which subset of passwords to save and only those will be
passed to the save method.

Bug: 1098672
Change-Id: I8d64005c190a1c9c3715fea25f004457dc6dff84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2288560
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#786681}