Remove SessionMonitor, add ReceiverSetupQuerier
These logs are sent to infrastructure that is turned down and no longer
used, so the SessionMonitor and its related code are deleted.

However, the SessionMonitor also included some features to query the
receiver for its setup information, including build version and the
receiver's friendly name. This information is used by the session in
order to apply hacks for different Chromecast versions.

This patch adds a new ReceiverSetupQuerier class that contains the setup
information code, as well as an updated and expanded unit test
class for the querier.

See https://chromium-review.googlesource.com/c/chromium/src/+/2240910.

Bug: 1015471
Change-Id: I700ae27e3241e4d0e9a997d5fbeee43d011dd4e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2261394
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Reviewed-by: Yuri Wiitala <miu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786848}