[Media Feeds] Remove watchtime min for display
At the moment we require 30 mins of watchtime
for a feed to be fetched or displayed. This removes
the display requirement so devs can manually fetch
feeds and for them to be shown in the UI. For most
users the fetching will apply the watchtime
requirement.

BUG=1103749

Change-Id: I01d27f5a36f526a1c6bc26a991aa5ca3454826e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2290310
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Commit-Queue: Becca Hughes <beccahughes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786882}