[Switch Access] Point Scan infrastructure
This change sets up the infrastructure for future development of an
alternate form of selection via Switch Access called Point Scanning.
(go/cros-switch-point-scanning has more details).

Specifically, this change creates a PointScanController to drive the
logic and a PointScanLayer to draw the UI, as well as a command-line
to develop behind. Currently, if the flag
kEnableSwitchAccessPointScanning is enabled, Switch Access draws a
vertical line through the center of the screen when the feature is
turned on.

Bug: 1061537
Change-Id: Ifba390904d5e94ee5edbe494a6799168a95a7272
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2092731
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786903}